register controller debugging

+++ src/utils/cloudinary.js
@@ -4,7 +4,7 @@ import fs from "fs"
 cloudinary.config({
     cloud_name : process.env.CLOUDINARY_CLOUD_NAME,
     api_key : process.env.CLOUDINARY_API_KEY,
-    api_secret : process.env.CLOUDINARY_API_SECRET
+    api_secret : process.env.CLOUDINARY_API_SECRET,
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
@@ -16,10 +16,12 @@ const uploadOnCloudinary = async (localFilePath) => {
         })
         //File has been uploaded successfully
         console.log("File is uploaded on cloudinary");
-        console.log(response.url)
+        fs.unlinkSync(localFilePath);
         return response
     } catch (error) {
-        fs.unlinkSync(localFilePath)
+        if(fs.unlinkSync(localFilePath)){
+            fs.unlinkSync(localFilePath);
+        }
         return null;
     }
 }
